(pkg re) Renaming entry-point from interface.* to excmd.*
